requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# OCR & Document Processing
pytesseract
easyocr
passporteye
pdf2image
pdfplumber
pypdf
Pillow
python-dateutil
opencv-python

# Data Validation
pydantic
pydantic-settings

# Browser Automation
playwright

# LLM Clients (optional - for enhanced extraction)
openai
anthropic

# Utilities
python-dotenv
aiofiles
typing_extensions

# Agentic Framework
langgraph
langchain-core

# Development
pytest
pytest-asyncio
httpx
